backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.0
alpaca-py==0.20.2
pandas==2.1.3
pandas-ta==0.3.14b0
pytz==2023.3
aiohttp==3.9.1
setuptools==69.0.3
